fix job rate limiter
