+++ server/odc-service/src/main/java/com/oceanbase/odc/service/task/config/DefaultSpringJobConfiguration.java
@@ -31,7 +31,9 @@ import com.oceanbase.odc.service.schedule.ScheduleTaskService;
 import com.oceanbase.odc.service.task.TaskService;
 import com.oceanbase.odc.service.task.caller.K8sJobClient;
 import com.oceanbase.odc.service.task.dispatch.ImmediateJobDispatcher;
-import com.oceanbase.odc.service.task.schedule.MonitorExecutorStatusRateLimiter;
+import com.oceanbase.odc.service.task.schedule.MonitorProcessRateLimiter;
+import com.oceanbase.odc.service.task.schedule.StartJobRateLimiter;
+import com.oceanbase.odc.service.task.schedule.StartJobRateLimiterSupport;
 import com.oceanbase.odc.service.task.schedule.provider.DefaultHostUrlProvider;
 import com.oceanbase.odc.service.task.schedule.provider.DefaultJobImageNameProvider;
 import com.oceanbase.odc.service.task.service.SpringTransactionManager;
@@ -67,7 +69,7 @@ public class DefaultSpringJobConfiguration extends DefaultJobConfiguration
         setTaskFrameworkService(tfs);
         setEventPublisher(publisher);
         setTransactionManager(new SpringTransactionManager(ctx.getBean(TransactionTemplate.class)));
-        setStartJobRateLimiter(new MonitorExecutorStatusRateLimiter(this::getTaskFrameworkProperties, tfs));
+        initJobRateLimiter(tfs);
     }
 
     @Override
@@ -83,5 +85,13 @@ public class DefaultSpringJobConfiguration extends DefaultJobConfiguration
     @Override
     public K8sJobClient getK8sJobClient() {
         return ctx.getBean(K8sJobClient.class);
+    }
+
+    private void initJobRateLimiter(TaskFrameworkService tfs) {
+        StartJobRateLimiterSupport limiterSupport = new StartJobRateLimiterSupport();
+        limiterSupport.addJobRateLimiter(
+            new MonitorProcessRateLimiter(this::getTaskFrameworkProperties, tfs));
+        ctx.getBeansOfType(StartJobRateLimiter.class).forEach((k, v) -> limiterSupport.addJobRateLimiter(v));
+        setStartJobRateLimiter(limiterSupport);
     }
 }
