+++ server/odc-service/src/main/java/com/oceanbase/odc/service/task/config/K8sProperties.java
@@ -68,5 +68,9 @@ public class K8sProperties {
      */
     private Long mountDiskSize;
 
+    /**
+     * node count in pool
+     */
+    private Long nodeCount;
 
 }
