+++ server/odc-service/src/main/java/com/oceanbase/odc/service/task/schedule/StartJobRateLimiterSupport.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2024 OceanBase.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.oceanbase.odc.service.task.schedule;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author yaobin
+ * @date 2024-03-28
+ * @since 4.2.4
+ */
+public class StartJobRateLimiterSupport implements StartJobRateLimiter {
+    private final List<StartJobRateLimiter> limiters = new ArrayList<>();
+
+    @Override
+    public boolean tryAcquire() {
+        boolean result = true;
+        for (StartJobRateLimiter limiter : limiters) {
+            if (!limiter.tryAcquire()) {
+                result = false;
+                break;
+            }
+        }
+        return result;
+    }
+
+    public void addJobRateLimiter(StartJobRateLimiter limiter) {
+        limiters.add(limiter);
+    }
+
+}
